Refactor internal stuff into new internals/ directory.
