+++ api/stars_api.py
@@ -20,7 +20,7 @@ import logging
 
 from framework import basehandlers
 import models
-import notifier
+from internals import notifier
 
 
 class StarsAPI(basehandlers.APIHandler):
