+++ api/stars_api_test.py
@@ -25,7 +25,7 @@ import werkzeug.exceptions  # Flask HTTP stuff.
 from api import register
 from api import stars_api
 import models
-import notifier
+from internals import notifier
